requirements.txt
bottle>=0.12
requests>=2.31
pillow>=10.0
cairosvg>=2.7
